+++ src/ECommerce.Domain/Ordering/Entities/OrderItem.cs
@@ -8,12 +8,12 @@ public class OrderItem : BaseEntity
     public Guid ProductId { get; private set; }
     public string ProductName { get; private set; } = default!;
     public Money UnitPrice { get; private set; } = default!;
-    public int Quantity { get; private set; }
+    public decimal Quantity { get; private set; }
     public Money LineTotal => UnitPrice.Multiply(Quantity);
 
     private OrderItem() { }
 
-    public static OrderItem Create(Guid productId, string productName, Money unitPrice, int quantity)
+    public static OrderItem Create(Guid productId, string productName, Money unitPrice, decimal quantity)
     {
         if (quantity <= 0) throw new ArgumentException("Quantity must be positive.");
         return new OrderItem
